requirements.txt
asgiref==3.6.0
attrs==22.2.0
btrfsutil==6.1.3
cairocffi==1.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
configobj==5.1.0.dev0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cupshelpers==1.0
defusedxml==0.7.1
Django==4.1.7
django-guardian==2.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
greenlet==2.0.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
libfdt==1.6.1
MarkupSafe==2.1.2
msgpack==1.0.4
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
ply==3.11
psutil==5.9.4
psycopg2-binary==2.9.5
pycairo==1.23.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.2
PyGObject==3.42.2
PyJWT==2.6.0
pynvim==0.4.3
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
qtile==0.22.1
qutebrowser==2.5.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
team==1.0
terminator==2.1.2
uritemplate==4.1.1
urllib3==1.26.14
validate==5.1.0.dev0
xcffib==1.2.0
